+++ lib/src/chat_theme.dart
@@ -43,7 +43,7 @@ const SECONDARY = Color(0xfff5f5f7);
 /// Secondary dark
 const SECONDARY_DARK = Color(0xff2b2250);
 
-/// Base chat theme containing all required variables to make a theme.
+/// Base chat theme containing all required properties to make a theme.
 /// Extend this class if you want to create a custom theme.
 @immutable
 abstract class ChatTheme {
@@ -59,8 +59,11 @@ abstract class ChatTheme {
     required this.errorIcon,
     required this.inputBackgroundColor,
     required this.inputBorderRadius,
-    required this.inputTextStyle,
+    required this.inputPadding,
     required this.inputTextColor,
+    this.inputTextCursorColor,
+    required this.inputTextDecoration,
+    required this.inputTextStyle,
     required this.messageBorderRadius,
     required this.primaryColor,
     required this.receivedMessageBodyTextStyle,
@@ -113,8 +116,17 @@ abstract class ChatTheme {
   /// Top border radius of the bottom bar where text field is
   final BorderRadius inputBorderRadius;
 
+  /// Insets of the bottom bar where text field is
+  final EdgeInsetsGeometry inputPadding;
+
   /// Color of the text field's text and attachment/send buttons
   final Color inputTextColor;
+
+  /// Color of the text field's cursor
+  final Color? inputTextCursorColor;
+
+  /// Decoration of the input text field
+  final InputDecoration inputTextDecoration;
 
   /// Text style of the message input. To change the color use [inputTextColor].
   final TextStyle inputTextStyle;
@@ -195,7 +207,7 @@ abstract class ChatTheme {
 @immutable
 class DefaultChatTheme extends ChatTheme {
   /// Creates a default chat theme. Use this constructor if you want to
-  /// override only a couple of variables, otherwise create a new class
+  /// override only a couple of properties, otherwise create a new class
   /// which extends [ChatTheme]
   const DefaultChatTheme({
     Widget? attachmentButtonIcon,
@@ -222,7 +234,14 @@ class DefaultChatTheme extends ChatTheme {
     BorderRadius inputBorderRadius = const BorderRadius.vertical(
       top: Radius.circular(20),
     ),
+    EdgeInsetsGeometry inputPadding = EdgeInsets.zero,
     Color inputTextColor = NEUTRAL_7,
+    Color? inputTextCursorColor,
+    InputDecoration inputTextDecoration = const InputDecoration(
+      border: InputBorder.none,
+      contentPadding: EdgeInsets.zero,
+      isCollapsed: true,
+    ),
     TextStyle inputTextStyle = const TextStyle(
       fontFamily: 'Avenir',
       fontSize: 16,
@@ -319,7 +338,10 @@ class DefaultChatTheme extends ChatTheme {
           errorIcon: errorIcon,
           inputBackgroundColor: inputBackgroundColor,
           inputBorderRadius: inputBorderRadius,
+          inputPadding: inputPadding,
           inputTextColor: inputTextColor,
+          inputTextCursorColor: inputTextCursorColor,
+          inputTextDecoration: inputTextDecoration,
           inputTextStyle: inputTextStyle,
           messageBorderRadius: messageBorderRadius,
           primaryColor: primaryColor,
@@ -350,7 +372,7 @@ class DefaultChatTheme extends ChatTheme {
 @immutable
 class DarkChatTheme extends ChatTheme {
   /// Creates a dark chat theme. Use this constructor if you want to
-  /// override only a couple of variables, otherwise create a new class
+  /// override only a couple of properties, otherwise create a new class
   /// which extends [ChatTheme]
   const DarkChatTheme({
     Widget? attachmentButtonIcon,
@@ -377,7 +399,14 @@ class DarkChatTheme extends ChatTheme {
     BorderRadius inputBorderRadius = const BorderRadius.vertical(
       top: Radius.circular(20),
     ),
+    EdgeInsetsGeometry inputPadding = EdgeInsets.zero,
     Color inputTextColor = NEUTRAL_7,
+    Color? inputTextCursorColor,
+    InputDecoration inputTextDecoration = const InputDecoration(
+      border: InputBorder.none,
+      contentPadding: EdgeInsets.zero,
+      isCollapsed: true,
+    ),
     TextStyle inputTextStyle = const TextStyle(
       fontFamily: 'Avenir',
       fontSize: 16,
@@ -474,7 +503,10 @@ class DarkChatTheme extends ChatTheme {
           errorIcon: errorIcon,
           inputBackgroundColor: inputBackgroundColor,
           inputBorderRadius: inputBorderRadius,
+          inputPadding: inputPadding,
           inputTextColor: inputTextColor,
+          inputTextCursorColor: inputTextCursorColor,
+          inputTextDecoration: inputTextDecoration,
           inputTextStyle: inputTextStyle,
           messageBorderRadius: messageBorderRadius,
           primaryColor: primaryColor,