+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 1.3.2
+
+- Fix memory leak. Thanks @m-j-g for reporting!
+- Add `customBottomWidget` useful to remove the input and create a channel view
+- Add `inputPadding`, `inputTextCursorColor` and `inputTextDecoration` to the theme for the additional input customization
+
+## 1.3.1
+
+- Rename `buildCustomMessage` to `customMessageBuilder`
+- Update dependencies
+
 ## 1.3.0
 
 - Chat is now correctly rendered in not full screen mode. Removed `SafeArea` from the lib itself, wrap `Chat` component if needed. Thanks @m-j-g for reporting!